+++ es/core/computed/init-module-computed.js
@@ -5,7 +5,8 @@ import { CATE_MODULE } from '../../support/constant';
 import configureDepFns from '../base/configure-dep-fns';
 import pickDepFns from '../base/pick-dep-fns';
 
-const { safeGetObjectFromObject, isPlainJsonObject } = util;
+const { safeGetObjectFromObject, isPlainJsonObject, makeCommitHandler } = util;
+const callInfo = { payload: null, renderKey: '', delay: -1 };
 
 export default function (module, computed, append = false, configureDep = true) {
   if (!isPlainJsonObject(computed)) {
@@ -33,13 +34,19 @@ export default function (module, computed, append = false, configureDep = true) 
   }
 
   const { pickedFns, setted, changed } = pickDepFns(true, CATE_MODULE, 'computed', rootComputedDep, module, moduleState, moduleState);
-  pickedFns.forEach(({ retKey, fn, depKeys }) => {
-    const fnCtx = { retKey, isBeforeMount:false, setted, changed, stateModule: module, refModule: null, oldState: moduleState, committedState: moduleState, refCtx: null };
 
-    const computedValue = util.executeCompOrWatch(retKey, depKeys, fn, moduleState, moduleState, fnCtx);
+  if (pickedFns.length ) {
+    const d = ccContext.getDispatcher();
+    const { commit, flush } = makeCommitHandler(module, d && d.ctx.changeState, callInfo);
 
-    const moduleComputedValue = safeGetObjectFromObject(rootComputedValue, module);
-    moduleComputedValue[retKey] = computedValue;
-  });
+    pickedFns.forEach(({ retKey, fn, depKeys }) => {
+      const fnCtx = { retKey, payload: null, isFirstCall: true, commit, setted, changed, stateModule: module, refModule: null, oldState: moduleState, committedState: moduleState, refCtx: null };
+      const computedValue = util.executeCompOrWatch(retKey, depKeys, fn, moduleState, moduleState, fnCtx);
+      const moduleComputedValue = safeGetObjectFromObject(rootComputedValue, module);
+      moduleComputedValue[retKey] = computedValue;
+    });
+
+    flush();
+  }
 
 }